requirements.txt
boto3==1.34.85
botocore==1.34.85
certifi==2024.2.2
charset-normalizer==3.3.2
confluent-kafka==2.3.0
idna==3.7
jmespath==1.0.1
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
